backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.6
google-adk>=0.3.0
google-cloud-aiplatform>=1.67.1
google-generativeai>=0.8.3
google-auth>=2.35.0
google-auth-oauthlib>=1.2.1
google-api-python-client>=2.149.0
asyncpg>=0.29.0
sqlalchemy[asyncio]>=2.0.36
alembic>=1.13.3
pgvector>=0.3.5
pydantic>=2.10.0
pydantic-settings>=2.6.0
python-jose[cryptography]>=3.3.0
httpx>=0.27.2
google-cloud-pubsub>=2.27.1
python-multipart>=0.0.12
psycopg2-binary>=2.9.0
cloud-sql-python-connector[asyncpg]>=1.10.0
